// frontend/src/api/gen/schemas/teamBlogSection.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * hackathon-portal-api
 * OpenAPI spec version: 1.0.0
 */
import type { BlogSectionLayout } from "./blogSectionLayout";

export interface TeamBlogSection {
  content: string;
  /** @nullable */
  image_id?: string | null;
  /** @nullable */
  image_url?: string | null;
  layout: BlogSectionLayout;
}